\name{FunChisq-deprecated}
\alias{FunChisq-deprecated}
\alias{cp.chisq.test}

\title{Deprecated Functions in Package \pkg{FunChisq}}

\description{
  These functions are provided for compatibility with older versions of package \pkg{FunChisq} only, and may be removed eventually.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{\code{cp.chisq.test}: now available as \code{\link[DiffXTables]{cp.chisq.test}}} in package \pkg{DiffXTables}

  }
}
