\name{FunChisq-package}
\alias{FunChisq-package}
\docType{package}
\title{
Nonparametric functional dependency tests based on chi-squares
}
\description{
The package implements non-constant functional chi-square tests (Zhang and Song, 2013). Such tests offer an advantage over regression analysis when a parametric form cannot be reliably assumed for the underlying function. It requires data from two or more variables be formatted as a contingency table. Continuous variables must be discretized first which can be done by using R package Ckmeans.1d.dp.
}
\details{
\tabular{ll}{
Package: \tab FunChisq\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-05\cr
License: \tab  LGPL (>= 3) \cr
}

}
\seealso{

  The \code{Ckmeans.1d.dp} package.
}

\author{

Yang Zhang and Joe Song. 

}
\references{

Zhang, Y. and Song, M. (2013) \emph{Deciphering interactions in causal networks without parametric assumptions.} 
\url{http://arxiv.org/abs/1311.2707}

}

\keyword{ package }
